+++ core/src/main/java/cn/leancloud/LCLeaderboard.java
@@ -119,7 +119,7 @@ public class LCLeaderboard {
         }
     }
 
-    void setVersion(int version) {
+    public void setVersion(int version) {
         this.version = version;
     }