+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 Following is change logs for recently release versions, you can refer to [releases page](https://github.com/leancloud/java-unified-sdk/releases) for more details.
 
+## 8.1.0 release
+
+#### Break changes
+- None
+
+#### New features
+- add leaderboard APIs.
+- add LCFile#setPathPrefix(String prefix) method, enable save file with specified path.
+- add LCIMConversationsQuery#whereContainedIn method.
+
+#### Optimization and fixed bugs
+- None.
+
 ## 8.0.4 release
 
 #### Break changes